'replace in'->'replace' for expr replacement
It is my fallible understanding that for `kibit.replace/replace-expr`,
the return value is the entire expression.

Leaving aside the discussion that `replace-file` might perhaps be more
optimally rendered as `replace-file!`—and `replace-expr` as
`convert-expr` or similar—what seems clear is that the entire
expression is being operated on and returned.

The replacement operation is not being done only in part. Therefore
terminal preposition `in` misses the mark